# Make the lease refresh period configurable