### description

This PR proposes to render configurable the `5s` value hard-coded here:
https://github.com/vmware/vmware-go-kcl/blob/eca3d01e05738f915b32567ea702996539a6b850/clientlibrary/worker/shard-consumer.go#L172

I called this parameter `LeaseRefreshPeriodMillis` and its default value is 5s.

### rationale behind this change

For our specific use case we control the read throughput by performing a `time.Sleep` in the `ProcessRecords` method of our `RecordProcessor` implementation. This has the benefit to sensibly decrease the CPU usage (by performing less IO syscalls) while keeping the same average read throughput we would obtain by decreasing `MaxRecords`.

However for this use case the 5s value hard-coded here is not sufficient.

If the time taken by `GetRecords+ProcessRecords` exceed 5s then there is a problem:
 - current shard owner believes it will still own the shard 5 seconds later but, 
 - in fact this is not true, if `GetRecords+ProcessRecords` exceeds 5s then the lease duration will be reached and the shard might be offered to, and taken by, another worker.

This leads to records duplication because both the old and new shard owners will read from the current checkpoint until the lease end. In our use case this led to 4% record duplication.

On this screenshot showing the shard count for a stream with 20 shards, and 2 workers. With this fix the number of shards per worker stays constant at 10, while before the problem led to a never ending variation in the number of shards per worker. 
![image](https://user-images.githubusercontent.com/476650/68759882-dd48ca80-0610-11ea-8bc6-fbb4b6d94be0.png)

I reckon this behavior is amplified in our use case because `ProcessRecords` always takes more time than 5s but it could anyway happen in other cases: as the time taken by `GetRecords+ProcessRecords` is never guaranteed to be below 5s (due to IO delay for example).


### another, more correct solution

There would be another way to solve this: we could guarantee that `GetRecords+ProcessRecords` never exceeds a certain -configurable- amount of time. By using a [`context.WithDeadline`](https://golang.org/pkg/context/#WithDeadline) and passing it around, using [`GetRecordsWithContext`](https://github.com/aws/aws-sdk-go/blob/v1.25.30/service/kinesis/api.go#L1339) instead of `GetRecords`, and having the Checkpointer take a context as well (to forward it to [`PutItemWithContext`](https://docs.aws.amazon.com/sdk-for-go/api/service/dynamodb/#DynamoDB.PutItemWithContext) instead of `PutItem`.

But this would be a breaking change as it would change many of the exported interfaces.